# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "MetaShare API",
    "description": "MetaShare 相關 API"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "JSON Web Token": {
      "type": "apiKey",
      "schema": "bearer",
      "bearerFormat": "JWT",
      "in": "headers",
      "name": "authorization",
      "description": "Bearer token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
    "/api/posts": {
      "get": {
        "tags": [
          "動態貼文"
        ],
        "summary": "取得所有貼文 API",
        "description": "",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "關鍵字",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "排序方式，desc 為新至舊，asc 為舊至新",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "取得所有貼文成功",
            "schema": {
              "$ref": "#/definitions/getAllPosts_Schema"
            }
          },
          "401": {
            "description": "未登入狀態",
            "schema": {
              "$ref": "#/definitions/error_noLoggin_Schema"
            }
          }
        },
        "security": [
          {
            "JSON Web Token": []
          }
        ]
      }
    },
...
    "/api/post/{id}/likes": {
      "post": {
        "tags": [
          "動態貼文"
        ],
        "summary": "新增貼文讚 API",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "輸入 Post id",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "按讚成功",
            "schema": {
              "$ref": "#/definitions/postLike_Schema"
            }
          },
          "400": {
            "description": "找不到 ID",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Error"
                },
                "message": {
                  "type": "string",
                  "example": "找不到 id，請重新確認"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "無效的 ID",
            "schema": {
              "$ref": "#/definitions/error_postId_Schema"
            }
          }
        },
        "security": [
          {
            "JSON Web Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "動態貼文"
        ],
        "summary": "取消增貼文讚 API",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "輸入 Post id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "取消按讚成功",
            "schema": {
              "$ref": "#/definitions/postLike_Schema"
            }
          },
          "400": {
            "description": "找不到 ID",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Error"
                },
                "message": {
                  "type": "string",
                  "example": "找不到 id，請重新確認"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "無效的 ID",
            "schema": {
              "$ref": "#/definitions/error_postId_Schema"
            }
          }
        },
        "security": [
          {
            "JSON Web Token": []
          }
        ]
      }
    },
    "/api/post/{id}/comments": {
      "post": {
        "tags": [
          "動態貼文"
        ],
        "summary": "新增貼文留言 API",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "輸入 Post id",
            "type": "string"
          },
          {
            "name": "parameters_name",
            "in": "body",
            "description": "新增留言的格式",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "type": "string",
                  "example": "你好"
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "新增留言成功",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "object",
                      "properties": {
                        "comment": {
                          "type": "string",
                          "example": "嘿嘿"
                        },
                        "userId": {
                          "type": "string",
                          "example": "62ee6d4fdb480a3990f5d3ee"
                        },
                        "postId": {
                          "type": "string",
                          "example": "63993f9282069f3f28815d92"
                        },
                        "_id": {
                          "type": "string",
                          "example": "639a85c14f4144f93c8b49fb"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2022-12-15T02:26:09.634Z"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "無法找到此貼文",
            "schema": {
              "$ref": "#/definitions/error_post_Schema"
            }
          }
        },
        "security": [
          {
            "JSON Web Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "動態貼文"
        ],
        "summary": "刪除貼文留言 API",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "輸入 Comment id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "刪除留言成功",
            "schema": {
              "$ref": "#/definitions/commentPost_Schema"
            }
          },
          "400": {
            "description": "找不到 ID",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Error"
                },
                "message": {
                  "type": "string",
                  "example": "找不到 id，請重新確認"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "無效的 ID",
            "schema": {
              "$ref": "#/definitions/error_postId_Schema"
            }
          }
        },
        "security": [
          {
            "JSON Web Token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "動態貼文"
        ],
        "summary": "編輯貼文留言 API",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "輸入 Comment id",
            "type": "string"
          },
          {
            "name": "parameters_name",
            "in": "body",
            "description": "編輯留言的格式",
            "schema": {
              "type": "object",
              "properties": {
                "comment": {
                  "type": "string",
                  "example": "你好"
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "編輯留言成功",
            "schema": {
              "$ref": "#/definitions/commentPost_Schema"
            }
          },
          "400": {
            "description": "找不到 ID",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Error"
                },
                "message": {
                  "type": "string",
                  "example": "找不到 id，請重新確認"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "無效的 ID",
            "schema": {
              "$ref": "#/definitions/error_postId_Schema"
            }
          }
        },
        "security": [
          {
            "JSON Web Token": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "error_noLoggin_Schema": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "Error"
        },
        "message": {
          "type": "string",
          "example": "您尚未登入"
        }
      }
    },
    "error_postId_Schema": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "Error"
        },
        "message": {
          "type": "string",
          "example": "無效的 ID，請重新確認"
        }
      }
    },
    "error_post_Schema": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "Error"
        },
        "message": {
          "type": "string",
          "example": "無法找到此貼文，請重新確認"
        }
      }
    },
...
    "getAllPosts_Schema": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "example": true
        },
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "63866de7430a8841e0be9744"
              },
              "user": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "example": "63866d0c430a8841e0be9676"
                  },
                  "name": {
                    "type": "string",
                    "example": "Test"
                  },
                  "avatar": {
                    "type": "string",
                    "example": "https://i.imgur.com/erdqB0d.jpg"
                  },
                  "role": {
                    "type": "string",
                    "example": "user"
                  },
                  "personalInfo": {
                    "type": "string",
                    "example": "測試"
                  }
                }
              },
              "content": {
                "type": "string",
                "example": "嘿嘿"
              },
              "image": {
                "type": "string",
                "example": ""
              },
              "likes": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "63866d0c430a8841e0be9676"
                    },
                    "name": {
                      "type": "string",
                      "example": "Test"
                    },
                    "avatar": {
                      "type": "string",
                      "example": "https://i.imgur.com/erdqB0d.jpg"
                    },
                    "role": {
                      "type": "string",
                      "example": "user"
                    }
                  }
                }
              },
              "createdAt": {
                "type": "string",
                "example": "2022-11-29T20:39:03.174Z"
              },
              "comments": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "638c480733ecc5b9b42c3715"
                    },
                    "comment": {
                      "type": "string",
                      "example": "嘿"
                    },
                    "userId": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "62ed1e72671067b1754e77cd"
                        },
                        "name": {
                          "type": "string",
                          "example": "小宥"
                        },
                        "avatar": {
                          "type": "string",
                          "example": "https://i.imgur.com/Ls0oWis.jpg"
                        },
                        "role": {
                          "type": "string",
                          "example": "user"
                        },
                        "personalInfo": {
                          "type": "string",
                          "example": "終於成功!!!!"
                        },
                        "passwordResetExpires": {
                          "type": "string",
                          "example": "2022-09-29T22:58:14.425Z"
                        }
                      }
                    },
                    "postId": {
                      "type": "string",
                      "example": "63866de7430a8841e0be9744"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2022-12-04T07:11:03.819Z"
                    }
                  }
                }
              },
              "id": {
                "type": "string",
                "example": "63866de7430a8841e0be9744"
              }
            }
          }
        }
      }
    },
...
    "commentPost_Schema": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "639a911fda4373b7ac2a4d37"
            },
            "comment": {
              "type": "string",
              "example": "你好"
            },
            "userId": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "62ee6d4fdb480a3990f5d3ee"
                },
                "name": {
                  "type": "string",
                  "example": "柚子"
                },
                "avatar": {
                  "type": "string",
                  "example": "https://i.imgur.com/ImgTFX5.jpg"
                },
                "role": {
                  "type": "string",
                  "example": "user"
                },
                "personalInfo": {
                  "type": "string",
                  "example": "我是一個人"
                }
              }
            },
            "postId": {
              "type": "string",
              "example": "63993f9282069f3f28815d92"
            },
            "createdAt": {
              "type": "string",
              "example": "2022-12-15T03:14:39.770Z"
            }
          }
        }
      }
    },
    "postLike_Schema": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "postId": {
          "type": "string",
          "example": "63993f9282069f3f28815d92"
        },
        "userId": {
          "type": "string",
          "example": "62ee6d4fdb480a3990f5d3ee"
        }
      }
    },
...
  }
}